requirements.txt
setuptools
wheel
spacy
lemminflect
pychalk